// src/app/home/models/tb-unit.ts
export class TbUnit {
  unitSerial?: number= null;
  unitCode?: string= null;
  unitNameEng?: string= null;
  unitNameTh?: string= null;
  unitPkg?: string= null;

  unitPpl?: string= null;
  weightFlag?: string= null;
  
  weightDesc?: string= null;

  offset?: number= null;
  limit?: number= null;
  
  
  createBy?: string= null;
  updateBy?: string= null;
}